rest-api/requirements.txt
Django==2.0
psycopg2==2.7.3.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djoser==1.1.5
factory_boy==2.9.2
django-silk==1.1.0
clearbit==0.1.7
email_hunter_python==1.1.0
punter==0.1.2
pyyaml==3.12
asyncio==3.4.3
aiohttp==2.3.9
pyotp==2.2.6
